synthmark: add series of utilization tests
The -ts option allow the user to request a series of utilization
benchmarks with increasing numbers of voices.
The results are printed in CSV format so that they can be easily
plotted.
The -N option can be used to set the upper limit. If not specified
then a VoiceMark95 will be run to determine the upper limit.

Test: adb shell synthmark -ts