Optimize LatencyMark search algorithm.
Refactor into a suite that run a glitch test multiple times.
The search algorithm is similar to a binary search but
tries to avoid running tests that do not glitch, because they take longer.
It is also unbounded at the top so it uses an exponentially increasing
size to find the initial bounds.