Remove portaudio binary
This will fix a security warning:
Project is out of compliance with Binary Artifacts policy: binaries present in source code

Portaudio was only used for the Mac build.
The Mac build is now broken but it was already unused and unsupported.
If we want to restore it then we should make portaudio a submodule.

Note that there was very little actual security risk because PortAudio
is my project, along with Ross Bencina, and we closely review
all modifications.